Move messages to index
